pyproject.toml
[tool.poetry]
name = "mailing-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mailing_service"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"


[tool.poetry.group.dev.dependencies]
pytils = "^0.4.1"
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
